requirements.txt
fireplace2==1.0.0
PySide6

tensorflow~=2.9.1

hearthstone~=5.56.0
numpy~=1.22.3
waiting